Crowlet model complete
I completed the crowlet model. This is the last 3D model I had in mind for the crowsnest room. Everything else will be 2D assets if needed. I also placed models in their own folders. So you may have to re-open the models in Unity to update. Everything is still there.